Move Section classes to their own folder
