+++ src/Angor/Avalonia/AngorApp/Sections/Shell/IMainViewModel.cs
@@ -1,3 +1,5 @@
+using AngorApp.Sections.Shell.Sections;
+
 namespace AngorApp.Sections.Shell;
 
 public interface IMainViewModel
