+++ src/Angor/Avalonia/AngorApp/Sections/Shell/Sections/CommandSection.cs
@@ -1,0 +1,10 @@
+using System.Windows.Input;
+
+namespace AngorApp.Sections.Shell.Sections;
+
+public class CommandSection(string name, ICommand command, string icon) : SectionBase
+{
+    public string Name { get; } = name;
+    public ICommand Command { get; } = command;
+    public string Icon { get; } = icon;
+}
