+++ src/Angor/Avalonia/AngorApp/Sections/Shell/Sections/IContentSection.cs
@@ -1,0 +1,8 @@
+namespace AngorApp.Sections.Shell.Sections;
+
+public interface IContentSection
+{
+    string Name { get; }
+    Func<object?> GetViewModel { get; }
+    object? Icon { get; }
+}
