+++ src/Angor/Avalonia/AngorApp/Sections/Shell/Sections/Separator.cs
@@ -1,0 +1,3 @@
+namespace AngorApp.Sections.Shell.Sections;
+
+public class Separator : SectionBase;
