--- src/Angor/Avalonia/AngorApp/Sections/Shell/Separator.cs
@@ -1,3 +1,0 @@
-namespace AngorApp.Sections.Shell;
-
-public class Separator : SectionBase;
